refactor: move constants to constants package
