+++ packages/application-shell-connectors/src/utils/http-client.spec.ts
@@ -2,7 +2,7 @@
 import { rest } from 'msw';
 import type { Headers } from 'headers-polyfill';
 import { setupServer } from 'msw/node';
-import { SUPPORTED_HEADERS } from '../constants';
+import { SUPPORTED_HEADERS } from '@commercetools-frontend/constants';
 import {
   buildApiUrl,
   createHttpClientOptions,
