+++ packages/application-shell-connectors/src/utils/http-client.ts
@@ -8,8 +8,8 @@ import omitEmpty from 'omit-empty-es';
 import {
   type ApplicationWindow,
   STATUS_CODES,
+  SUPPORTED_HEADERS,
 } from '@commercetools-frontend/constants';
-import { SUPPORTED_HEADERS } from '../constants';
 import getCorrelationId from './get-correlation-id';
 import getMcApiUrl from './get-mc-api-url';
 import * as oidcStorage from './oidc-storage';
